Do not crash on invalid instructions
Stop the current basic block on an invalid instruction. Abort if the
resulting basic block is empty.

It opens the door for handling self-modifying code, as 2 new tests cases
show it. First, in case of invalid instruction being overwritten, qbdi
handles it natively. Second, in case of valid instruction being
overwritten, it lets a user adding an callback to invalidate cache.