Add Jit information and reduce cache size
Add a new Analysis (ANALYSIS_JIT) that return the address where the
patch is JITTed. This analysis may be used in case of fault inside
Jitted code to find the patch (and the associated instruction) that
generate the fault. This also include an API to find the patch given a
JIT address.

This patch also include two API that may be used to reduce the cache
size of QBDI.
- getNbExecBlock : return the number of execblock in the cache
- reduceCacheTo : reduce the cache by removing oldest CodeRegion
  (usually 1 CodeRegion == 1 ExecBlock)
While QBDI doesn't provide a cache limit or a precise indication of
memory usage, these API can be used in BASIC_BLOCK_NEW VM Event to
reduce the cache size in order to limit the memory usage of QBDI.